Create `CreatePVC` and move `GetTargetVolume`
